+++ playground/src/index.tsx
@@ -3,6 +3,7 @@ import * as hubs from 'frog/hubs'
 import { devtools } from 'frog/dev'
 import { serveStatic } from '@hono/node-server/serve-static'
 
+import { app as fontsApp } from './fonts.js'
 import { app as middlewareApp } from './middleware.js'
 import { app as neynarApp } from './neynar.js'
 import { app as routingApp } from './routing.js'
@@ -220,6 +221,7 @@ export const app = new Frog({
       ],
     })
   })
+  .route('/fonts', fontsApp)
   .route('/middleware', middlewareApp)
   .route('/neynar', neynarApp)
   .route('/routing', routingApp)