+++ ui/src/lib/store.ts
@@ -1,6 +1,8 @@
+import lz from 'lz-string'
 import { createStore } from 'zustand/vanilla'
+import { subscribeWithSelector } from 'zustand/middleware'
 
-import { Data, User } from '../types/frog'
+import { Bootstrap, Data, User } from '../types/frog'
 
 export type State = {
   dataKey: string
@@ -20,7 +22,7 @@ export type State = {
   tab: 'context' | 'meta-tags' | 'request' | 'state'
 }
 
-const initialState = {
+const initialState: State = {
   dataKey: '',
   dataMap: {},
   frameUrls: [],
@@ -38,4 +40,82 @@ const initialState = {
   tab: 'request',
 } satisfies State
 
-export const store = createStore<State>(() => initialState)
+export const store = createStore(subscribeWithSelector(() => initialState))
+
+const hashKey = 'state'
+
+export function hydrateStore(bootstrap: Bootstrap) {
+  const { data, frameUrls, user } = bootstrap
+
+  let hydrated: Partial<State> = {
+    frameUrls,
+    user,
+  }
+  const restoredState = restoreState()
+  if (restoredState) hydrated = { ...hydrated, ...restoredState }
+  else if (data)
+    hydrated = {
+      ...hydrated,
+      dataKey: data.id,
+      dataMap: { [data.id]: data },
+      logs: [data.id],
+      stack: [data.id],
+      frameUrls,
+      user,
+    }
+
+  store.setState((state) => ({
+    ...state,
+    ...hydrated,
+    overrides: user
+      ? {
+          ...state.overrides,
+          userFid: user.userFid,
+        }
+      : state.overrides,
+  }))
+
+  watchState()
+}
+
+function restoreState() {
+  try {
+    console.debug('[frog] restoring state...')
+    const state = location.hash.replace(`#${hashKey}/`, '').trim()
+    let restored = lz.decompressFromEncodedURIComponent(state)
+    // Fallback incase there is an extra level of decoding:
+    // https://gitter.im/Microsoft/TypeScript?at=5dc478ab9c39821509ff189a
+    if (!restored)
+      restored = lz.decompressFromEncodedURIComponent(decodeURIComponent(state))
+
+    const parsed = JSON.parse(restored ?? 'null')
+    if (!parsed) return undefined
+
+    const logIndex =
+      parsed.logIndex === (parsed.logs?.length ?? 0) - 1 ? -1 : parsed.logIndex
+    console.debug('[frog] restored state.')
+    return { ...parsed, logIndex }
+  } catch (error) {
+    console.log(`[frog] failed to restore state (${(error as Error).message})`)
+    history.replaceState(null, '', location.pathname)
+  }
+}
+
+function watchState() {
+  store.subscribe(
+    (state) => ({
+      dataKey: state.dataKey,
+      dataMap: state.dataMap,
+      logs: state.logs,
+      logIndex: state.logIndex,
+      overrides: state.overrides,
+      stack: state.stack,
+      stackIndex: state.stackIndex,
+      tab: state.tab,
+    }),
+    (slice) => {
+      const compressed = lz.compressToEncodedURIComponent(JSON.stringify(slice))
+      window.history.replaceState(null, '', `#${hashKey}/${compressed}`)
+    },
+  )
+}